feat: implement StudyroomGroupsDialog for enhanced group management in sidebar
